+++ Sources/EpoxyCore/Model/Providers/TraitCollectionProviding.swift
@@ -1,6 +1,7 @@
 // Created by eric_horacek on 12/16/20.
 // Copyright © 2020 Airbnb Inc. All rights reserved.
 
+#if !os(macOS)
 import UIKit
 
 /// The capability of providing a `UITraitCollection` instance.
@@ -10,3 +11,4 @@ public protocol TraitCollectionProviding {
   /// The `UITraitCollection` instance provided by this type.
   var traitCollection: UITraitCollection { get }
 }
+#endif