+++ Tests/EpoxyTests/CollectionViewTests/CollectionViewSpec.swift
@@ -1,11 +1,12 @@
 // Created by eric_horacek on 1/8/21.
 // Copyright © 2021 Airbnb Inc. All rights reserved.
 
-import EpoxyCollectionView
 import EpoxyCore
 import Nimble
 import Quick
 import UIKit
+
+@testable import EpoxyCollectionView
 
 // swiftlint:disable implicitly_unwrapped_optional
 
@@ -22,6 +23,23 @@ final class CollectionViewSpec: QuickSpec {
     required init?(coder _: NSCoder) {
       fatalError("init(coder:) has not been implemented")
     }
+  }
+
+  var mockCell: CollectionViewCell {
+    let cell = CollectionViewCell(frame: .zero)
+    cell.itemPath = .init(
+      itemDataID: DefaultDataID.noneProvided,
+      section: .dataID(DefaultDataID.noneProvided))
+    return cell
+  }
+
+  var mockHeaderView: CollectionViewReusableView {
+    let cell = CollectionViewReusableView(frame: .zero)
+    cell.itemPath = .init(
+      elementKind: UICollectionView.elementKindSectionHeader,
+      itemDataID: DefaultDataID.noneProvided,
+      section: .dataID(DefaultDataID.noneProvided))
+    return cell
   }
 
   override func spec() {
@@ -85,7 +103,7 @@ final class CollectionViewSpec: QuickSpec {
           beforeEach {
             collectionView.delegate?.collectionView?(
               collectionView,
-              willDisplay: CollectionViewCell(),
+              willDisplay: self.mockCell,
               forItemAt: IndexPath(item: 0, section: 0))
           }
 
@@ -99,7 +117,7 @@ final class CollectionViewSpec: QuickSpec {
           beforeEach {
             collectionView.delegate?.collectionView?(
               collectionView,
-              didEndDisplaying: CollectionViewCell(),
+              didEndDisplaying: self.mockCell,
               forItemAt: IndexPath(item: 0, section: 0))
           }
 
@@ -157,7 +175,7 @@ final class CollectionViewSpec: QuickSpec {
           beforeEach {
             collectionView.delegate?.collectionView?(
               collectionView,
-              willDisplaySupplementaryView: CollectionViewReusableView(),
+              willDisplaySupplementaryView: self.mockHeaderView,
               forElementKind: UICollectionView.elementKindSectionHeader,
               at: IndexPath(item: 0, section: 0))
           }
@@ -172,7 +190,7 @@ final class CollectionViewSpec: QuickSpec {
           beforeEach {
             collectionView.delegate?.collectionView?(
               collectionView,
-              didEndDisplayingSupplementaryView: CollectionViewReusableView(),
+              didEndDisplayingSupplementaryView: self.mockHeaderView,
               forElementOfKind: UICollectionView.elementKindSectionHeader,
               at: IndexPath(item: 0, section: 0))
           }
@@ -220,7 +238,7 @@ final class CollectionViewSpec: QuickSpec {
             beforeEach {
               collectionView.delegate?.collectionView?(
                 collectionView,
-                willDisplay: CollectionViewCell(),
+                willDisplay: self.mockCell,
                 forItemAt: IndexPath(item: 0, section: 0))
             }
 
@@ -236,7 +254,7 @@ final class CollectionViewSpec: QuickSpec {
               beforeEach {
                 collectionView.delegate?.collectionView?(
                   collectionView,
-                  didEndDisplaying: CollectionViewCell(),
+                  didEndDisplaying: self.mockCell,
                   forItemAt: IndexPath(item: 0, section: 0))
               }
 
@@ -248,7 +266,7 @@ final class CollectionViewSpec: QuickSpec {
                 beforeEach {
                   collectionView.delegate?.collectionView?(
                     collectionView,
-                    willDisplay: CollectionViewCell(),
+                    willDisplay: self.mockCell,
                     forItemAt: IndexPath(item: 0, section: 0))
                 }
 
@@ -266,7 +284,7 @@ final class CollectionViewSpec: QuickSpec {
               beforeEach {
                 collectionView.delegate?.collectionView?(
                   collectionView,
-                  willDisplaySupplementaryView: CollectionViewReusableView(),
+                  willDisplaySupplementaryView: self.mockHeaderView,
                   forElementKind: UICollectionView.elementKindSectionHeader,
                   at: IndexPath(item: 0, section: 0))
               }
@@ -279,11 +297,11 @@ final class CollectionViewSpec: QuickSpec {
                 beforeEach {
                   collectionView.delegate?.collectionView?(
                     collectionView,
-                    didEndDisplaying: CollectionViewCell(),
+                    didEndDisplaying: self.mockCell,
                     forItemAt: IndexPath(item: 0, section: 0))
                   collectionView.delegate?.collectionView?(
                     collectionView,
-                    didEndDisplayingSupplementaryView: CollectionViewReusableView(),
+                    didEndDisplayingSupplementaryView: self.mockHeaderView,
                     forElementOfKind: UICollectionView.elementKindSectionHeader,
                     at: IndexPath(item: 0, section: 0))
                 }